Added Unit Tests
What's New: The Unit tests were expanded to include the different items that could  be queried. The program will make sure a new instance of ollama is created and will be removed after.
What's needed: Preferable the function used the query ollama should be moved out of the file where it will be shared with all the other programs.